requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
countryinfo==0.1.2
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
telebot==0.0.5
urllib3==2.3.0
Werkzeug==3.1.3
